refactor(doc): extracting methods and refactoring components
