+++ packages/doc/content/components/components/button/default-native.mdx
@@ -6,7 +6,7 @@ You can use all differente compounds like `<Button.Outline />` and `<Button.Text
 ### Usage
 
 ```javascript type=expo
-<Button>Gympass</Button>
+<Button>Find an activity</Button>
 ```
 
 ### Props
