+++ packages/doc/content/components/components/button/default-web.mdx
@@ -5,9 +5,9 @@ You can use all differente compounds like `<Button.Outline />` and `<Button.Text
 
 ### Usage
 
-<!-- ```javascript
+```javascript
 <Button>Find an activity</Button>
-``` -->
+```
 
 ### Props
